+++ R/addNeighborhoodCases.R
@@ -12,7 +12,7 @@
 #' @param vestry Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.
 #' @param weighted Logical. \code{TRUE} computes shortest walking path weighted by road length. \code{FALSE} computes shortest walking path in terms of the number of nodes.
 #' @param color Character. Use a single color for all paths. \code{NULL} uses neighborhood colors defined by \code{snowColors().}
-#' @param case.location Character. For \code{metric = "euclidean"}: "address" uses \code{ortho.proj}; "nominal" uses \code{fatalities}.
+#' @param case.location Character. "address" uses \code{fatalities}; "orthogonal" uses \code{ortho.proj}.
 #' @param alpha.level Numeric. Alpha level transparency for area plot: a value in [0, 1].
 #' @param multi.core Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.
 #' @export
@@ -28,7 +28,7 @@
 addNeighborhoodCases <- function(pump.subset = NULL, pump.select = NULL,
   metric = "walking", type = "stack.base", token = "point", text.size = 0.5,
   pch = 16, point.size = 0.5, vestry = FALSE, weighted = TRUE, color = NULL,
-  case.location = "nominal", alpha.level = 0.5, multi.core = TRUE) {
+  case.location = "address", alpha.level = 0.5, multi.core = TRUE) {
 
   if (metric %in% c("euclidean", "walking") == FALSE) {
     stop('metric must be "euclidean" or "walking".')
@@ -42,8 +42,8 @@ addNeighborhoodCases <- function(pump.subset = NULL, pump.select = NULL,
     stop('token must be "id" or "point".')
   }
 
-  if (case.location %in% c("address", "nominal") == FALSE) {
-    stop('case.location must be "address" or "nominal".')
+  if (case.location %in% c("address", "orthogonal") == FALSE) {
+    stop('case.location must be "address" or "orthogonal".')
   }
 
   cores <- multiCore(multi.core)
@@ -95,11 +95,11 @@ addNeighborhoodCases <- function(pump.subset = NULL, pump.select = NULL,
   nom.data <- cholera::fatalities
 
   if (case.location == "address") {
-    case.data <- addr.data
-    vars <- c("x.proj", "y.proj")
-  } else if (case.location == "nominal") {
     case.data <- nom.data
     vars <- c("x", "y")
+  } else if (case.location == "orthogonal") {
+    case.data <- addr.data
+    vars <- c("x.proj", "y.proj")
   } else stop("Invalid case.location!")
 
   invisible(lapply(selected.pumps, function(x) {